// APP_RNM/frontend/src/modul/komite/api.ts
export interface BarisInboxKomite {
  kasusId: string
  tglUpdate: string
  statusWork: string
  klaimId: string
  nomorKlaim: string
  tingkatBerjalan: number
  komiteLoop: number
  nilaiKlaim: string
  mataUang: string
  statusBaris: string
}